apps/rest/pyproject.toml
[tool.poetry]
name = "sourcing"
version = "0.1.0"
description = ""
authors = ["David Lopez <davidlopez.hellin@Outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.1"}
motor = "^3.4.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
cryptography = "^42.0.5"
authlib = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
